pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "A simple AI agent that can answer questions and perform tasks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["google-genai==1.12.1", "python-dotenv==1.1.0"]

[tool.uv]
index-strategy = "unsafe-best-match"
